bomb

+++ SecretAdmin/Program.cs
@@ -17,29 +17,29 @@ namespace SecretAdmin
         public static ScpServer Server { get; private set; }
         public static CommandHandler CommandHandler { get; private set; }
         public static ConfigManager ConfigManager { get; private set; }
-
+        public static Logger ProgramLogger { get; private set; }
+        
         static void Main(string[] args)
         {
             Console.Title = $"SecretAdmin [v{Version}]";
             AppDomain.CurrentDomain.ProcessExit += OnExit;
             
-            Log.Intro();
-            Console.ReadKey();
+            Paths.Load();
             
-            //ConfigManager = new ConfigManager();
+            ConfigManager = new ConfigManager();
             
             if (ProgramIntroduction.FirstTime)
                 ProgramIntroduction.ShowIntroduction();
             
-            Paths.Load();
+            ConfigManager.LoadConfig();
+            ProgramLogger = new Logger(Path.Combine(Paths.ProgramLogsFolder, $"{DateTime.Now:MM.dd.yyyy-hh.mm.ss}.log"));
             
-            //ConfigManager.LoadConfig();
-
-            CommandHandler = new CommandHandler();
+            Log.Intro();
             
             Server = new ScpServer(new ServerConfig());
             Server.Start();
 
+            CommandHandler = new CommandHandler();
             InputManager.Start();
         }
 
@@ -48,7 +48,7 @@ namespace SecretAdmin
             Console.ForegroundColor = ConsoleColor.Blue;
             Console.WriteLine("\nExit Detected. Killing game process.");
 
-            //if(ConfigManager.SecretAdminConfig.SafeShutdown)
+            if (ConfigManager.SecretAdminConfig.SafeShutdown)
                 Server?.Kill();
             
             Console.ForegroundColor = ConsoleColor.Cyan;
